### 2019-06-20 #1479 @usethedata
Two declines and a deferral on reviewers.  Going back to look for additional options.  Suggestions welcomed.
