### 2019-06-22 #1479 @ikuyamada
@usethedata Thank you for working on finding reviewers of my article! However, I don't understand what kind of suggestions that are needed here. Are there a list of any reviewer candidates of this journal or can I recommend any researchers that are capable of reviewing this paper?
