### 2019-06-22 #1479 @usethedata
@ikuyamada What I'm looking for is if you know someone, who does not have a conflict of interest with you (see our [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and who would be qualified to review the work.  We have a list of [people who've volunteered as reviewers](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822), but I've already been working through that lists, looking for people with NLP and Python expertise.
