### 2019-06-27 #1479 @usethedata
👋  @jkahn @linuxscout @alexhanna @ionlights We're looking for a couple of reviewers for a JOSS paper "Wikipedia2Vec: An Optimized Tool for Learning Embeddings of Words and Entities from Wikipedia".  Would any of you be willing to take this on?  Thanks much for considering.
