### 2019-07-08 #1479 @ionlights
> 👋 @ionlights Thoughts?

@usethedata i think i can give this a review. it's my first review with JOSS. 😅i've given the following a look-over, as that all i need to know – aside from domain-specific knowledge?
- https://joss.readthedocs.io/en/latest/
- https://joss.readthedocs.io/en/latest/review_criteria.html
- https://joss.readthedocs.io/en/latest/review_checklist.html

i'll most likely be able to give this a look during my flights later this week, so if someone else chimes in, just let me know.
