### 2019-07-08 #1479 @usethedata
@ionlights Thanks much.  I am looking for a second reviewer (and have a couple of options).  If I can get one in the next couple of days, I'll open up the review issue then.  If I'm still looking by Wednesday evening, I'll open the review issue with just one reviewer and add the second one manually.  When I open the review issue, you'll see a checklist there for you, and that's where there would be any discussion about the review.
