### 2019-07-11 #1558 @whedon
[REVIEW]: Wikipedia2Vec: An Optimized Tool for Learning Embeddings of Words and Entities from Wikipedia
**Submitting author:** @ikuyamada (<a href="http://orcid.org/0000-0002-7980-7661">Ikuya Yamada</a>)
**Repository:** <a href="https://github.com/wikipedia2vec/wikipedia2vec" target ="_blank">https://github.com/wikipedia2vec/wikipedia2vec</a>
**Version:** v1.0.3
