### 2019-07-24 #1558 @ionlights
hey, @ikuyamda, @usethedata

sorry i've been silent. i'm nearly done. i've done all of the checklist save for the functionality (validating performance, etc.)

all the other, un-marked sections have comments which i've been preparing. i'm hoping to have this done this weekend, though i'm not sure about the performance validation.

@usethedata are you looking for me to reproduce their results, in terms of analytical performance (accuracies, scores, etc.) as well as training time? (latter is kinda up to whatever hardware user's have, so that's a bit harder to validate.)
