### 2019-08-04 #1558 @usethedata
@ionlights Sorry to have been out of communication, and I really thought I'd responded to your question.  Don't worry about training time in any absolute sense.  Anything you can do in a reasonable amount of time that checks for qualitative agreement with their claims is appreciated.
