### 2019-09-07 46df81d
add module: and timedtext: to ignored namespaces (related to #37)
