### 2019-09-07 4b8a6ce
remove leading and trailing whitespaces when extracting titles from wikilinks #37
