### 2019-09-07 c8caf8b
remove leading colon from titles #36
