### 2019-09-09 df50e7a
initial commit of text classification example
