### 2019-09-09 81900ea
remove unused code
