+++ examples/text_classification/README.md
@@ -4,12 +4,14 @@
 
 ## Introduction
 
-This example contains the implementation of **Neural Attentive Bag-of-Entities Model** (NABoE), a state-of-the-art neural network model that performs text classification using a set of Wikipedia entities (*bag of entities*) as an input.
+This directory contains the implementation of **Neural Attentive Bag-of-Entities Model** (NABoE), a state-of-the-art neural network model that performs text classification using a set of Wikipedia entities (*bag of entities*) as an input.
+
 For each entity name in a document (e.g., *Apple*), this model first detects Wikipedia entities that may be referred to by this name (e.g., *Apple Inc.*, *Apple (food)*) using a simple dictionary-based entity detector, and then computes the feature vector of the document using the weighted average of the entity embeddings of the detected entities.
 The weights are computed using a neural attention mechanism that enables the model to focus on a small subset of the entities that are less ambiguous in meaning and more relevant to the document.
 Text classification is performed using a simple logistic regression layer with the feature vector as an input.
 
-The code is built upon Wikipedia2Vec and the embeddings in this model are initialized using Wikipedia2Vec pretrained embeddings.
+The code is built upon Wikipedia2Vec and PyTorch.
+Further, the embeddings in this model are initialized using 300 dimensional Wikipedia2Vec pretrained embeddings.
 See [the paper](https://arxiv.org/abs/1909.01259) for further details.
 
 ## Results
@@ -18,7 +20,7 @@ See [the paper](https://arxiv.org/abs/1909.01259) for further details.
 
 | Model | Accuracy | F1 |
 | --- | --- | --- |
-| NABoE | **87.8** | **87.2** |
+| NABoE | **87.7** | **87.1** |
 | Bag-of-Words SVM | 79.0 | 78.3 |
 | BoE ([Jin et al., 2016](https://www.ijcai.org/Proceedings/16/Papers/401.pdf)) | 83.1 | 82.7 |
 | SWEM ([Shen et al., 2018](https://arxiv.org/abs/1805.09843)) | 85.3 | 85.5 |
@@ -29,14 +31,14 @@ See [the paper](https://arxiv.org/abs/1909.01259) for further details.
 
 | Model | Accuracy | F1 |
 | --- | --- | --- |
-| NABoE | **97.6** | **92.1** |
+| NABoE | **97.6** | **92.5** |
 | Bag-of-Words SVM | 94.7 | 85.1 |
 | BoE ([Jin et al., 2016](https://www.ijcai.org/Proceedings/16/Papers/401.pdf)) | 96.5 | 88.6 |
 | SWEM ([Shen et al., 2018](https://arxiv.org/abs/1805.09843)) | 96.7 | 89.8 |
 | TextEnt ([Yamada et al., 2018](https://arxiv.org/abs/1806.02960)) | 96.7 | 91.0 |
 | TextGCN ([Yao et al., 2018](https://arxiv.org/abs/1809.05679)) | 97.1 | - |
 
-NOTE: The above results are slightly better than the ones reported in the original paper because we refactored the code and tuned hyper parameters.
+NOTE: The above results are slightly better than the ones reported in the original paper because we improved the implementation and tuned hypeparameters.
 
 ## Reproducing Results on 20 Newsgroups
 
@@ -49,32 +51,34 @@ Note that you need **Python 3.6+** to run this experiment.
 % pip install -r requirements.txt
 ```
 
-**Download entity detector:**
[…]
