+++ examples/text_classification/requirements.txt
@@ -16,6 +16,6 @@ scikit-learn==0.21.3
 scipy==1.3.1
 six==1.12.0               # via wikipedia2vec
 soupsieve==1.9.3          # via beautifulsoup4
-torch==1.2.0
+torch==1.1.0
 tqdm==4.35.0
 wikipedia2vec==1.0.4
