+++ examples/text_classification/requirements.in
@@ -4,6 +4,6 @@ joblib==0.13.2
 numpy==1.17.2
 scipy==1.3.1
 scikit-learn==0.21.3
-torch==1.2.0
+torch==1.1.0
 tqdm==4.35.0
 wikipedia2vec==1.0.4
