+++ examples/text_classification/README.md
@@ -4,14 +4,14 @@
 
 ## Introduction
 
-This directory contains the implementation of **Neural Attentive Bag-of-Entities Model** (NABoE), a state-of-the-art neural network model that performs text classification using a set of Wikipedia entities (*bag of entities*) as an input.
+This directory contains the implementation of **Neural Attentive Bag-of-Entities Model** (NABoE), a state-of-the-art neural network model that performs text classification using a set of Wikipedia entities (*bag of entities*) as input.
 
 For each entity name in a document (e.g., *Apple*), this model first detects Wikipedia entities that may be referred to by this name (e.g., *Apple Inc.*, *Apple (food)*) using a simple dictionary-based entity detector, and then computes the feature vector of the document using the weighted average of the entity embeddings of the detected entities.
 The weights are computed using a neural attention mechanism that enables the model to focus on a small subset of the entities that are less ambiguous in meaning and more relevant to the document.
-Text classification is performed using a linear layer with the feature vector as an input.
+Text classification is performed using a linear layer with the feature vector as input.
 
 The code is built upon Wikipedia2Vec and PyTorch.
-Further, the embeddings in this model are initialized using 300-dimensional Wikipedia2Vec pretrained embeddings.
+The embeddings in this model are initialized using Wikipedia2Vec pretrained embeddings.
 See [the paper](https://arxiv.org/abs/1909.01259) for further details.
 
 ## Results
@@ -38,7 +38,7 @@ See [the paper](https://arxiv.org/abs/1909.01259) for further details.
 | TextEnt ([Yamada et al., 2018](https://arxiv.org/abs/1806.02960)) | 96.7 | 91.0 |
 | TextGCN ([Yao et al., 2018](https://arxiv.org/abs/1809.05679)) | 97.1 | - |
 
-NOTE: The above results are slightly better than the ones reported in the original paper because we tuned hyperparameters and added dropout regularization.
+NOTE: The above results are slightly better than the ones reported in the original paper because we tuned hyperparameters and improved the model implementation (i.e., adding weight decay, learning rate warmup, and dropout regularization).
 
 ## Reproducing Results on 20 Newsgroups
 
@@ -73,9 +73,9 @@ Note that you need **Python 3.6+** to run this experiment.
 
 You can speed up the training by specifying *--use-gpu* option if your machine has a GPU.
 
-## Build Your Own Wikipedia2Vec Pretrained Embeddings / Entity Detector
+## Building Wikipedia2Vec Pretrained Embeddings and Entity Detector
 
-You can easily build your own Wikipedia2Vec pretrained embeddings and entity detector based on a Wikipedia dump file (*enwiki-DATE-pages-articles.xml.bz2*) available at [Wikimedia Downloads](https://dumps.wikimedia.org/enwiki/).
+You can easily build Wikipedia2Vec pretrained embeddings and entity detector used in the above experiment based on a Wikipedia dump file (*enwiki-DATE-pages-articles.xml.bz2*) available at [Wikimedia Downloads](https://dumps.wikimedia.org/enwiki/).
 
 **Train Wikipedia2Vec pretrained embeddings:**
 
@@ -90,15 +90,15 @@ You can easily build your own Wikipedia2Vec pretrained embeddings and entity det
 % python main.py build-entity-linker DUMP_DB_FILE ENTITY_DETECTOR_FILE
 ```
 
-## References
+## Reference
 
-If you refer this model in a scientific publication, please cite the following paper:
+If you refer this text classification model in a scientific publication, please cite the following paper:
 
 ```bibtex
 @article{yamada2019neural,
   title={Neural Attentive Bag-of-Entities Model for Text Classification},
[…]
