+++ docs/index.md
@@ -54,6 +54,8 @@ Ikuya Yamada, Hiroyuki Shindo, Hideaki Takeda, Yoshiyasu Takefuji, [Joint Learni
 
 The text classification model contained in [this example](https://github.com/wikipedia2vec/wikipedia2vec/tree/master/examples/text_classification) was proposed in the following paper:
 
+Ikuya Yamada, Hiroyuki Shindo, [Neural Attentive Bag-of-Entities Model for Text Classification](https://arxiv.org/abs/1909.01259).
+
 ```bibtex
 @article{yamada2019neural,
   title={Neural Attentive Bag-of-Entities Model for Text Classification},
