### 2019-09-09 4d7f50a
Add text classification example (#39)
* initial commit of text classification example

* remove unused code

* update README

* add dropout

* update README

* downgrade PyTorch

* minor fixes

* add description and reference of NABoE

* fix docs
