+++ examples/text_classification/entity_linker.py
@@ -1,0 +1,189 @@
+import logging
+import re
+from collections import defaultdict, Counter
+from contextlib import closing
+from multiprocessing.pool import Pool
+import joblib
+from marisa_trie import Trie, RecordTrie
+from tqdm import tqdm
+
+from data import normalize_text
+
+logger = logging.getLogger(__name__)
+
+
+class Mention(object):
+    __slots__ = ('title', 'text', 'start', 'end', 'link_count', 'total_link_count', 'doc_count')
+
+    def __init__(self, title, text, start, end, link_count, total_link_count, doc_count):
+        self.title = title
+        self.text = text
+        self.start = start
+        self.end = end
+        self.link_count = link_count
+        self.total_link_count = total_link_count
+        self.doc_count = doc_count
+
+    @property
+    def span(self):
+        return self.start, self.end
+
+    @property
+    def link_prob(self):
+        if self.doc_count > 0:
+            return min(1.0, self.total_link_count / self.doc_count)
+        else:
+            return 0.0
+
+    @property
+    def prior_prob(self):
+        if self.total_link_count > 0:
+            return min(1.0, self.link_count / self.total_link_count)
+        else:
+            return 0.0
+
+    def __repr__(self):
+        return f'<Mention {self.text} -> {self.title}>'
+
+
+_dump_db = _tokenizer = _max_mention_length = _name_trie = None
+
+
+class EntityLinker(object):
+    def __init__(self, data_file):
+        data = joblib.load(data_file)
+        self.title_trie = data['title_trie']
+        self.mention_trie = data['mention_trie']
+        self.data_trie = data['data_trie']
+        self.tokenizer = data['tokenizer']
+        self.max_mention_length = data['max_mention_length']
+
+    def detect_mentions(self, text):
+        tokens = self.tokenizer.tokenize(text)
+        end_offsets = frozenset(t.span[1] for t in tokens)
+
+        ret = []
+        cur = 0
+        for token in tokens:
+            start = token.span[0]
+            if cur > start:
+                continue
+
+            for prefix in sorted(self.mention_trie.prefixes(text[start:start + self.max_mention_length]), key=len,
+                                 reverse=True):
+                end = start + len(prefix)
+                if end in end_offsets:
+                    matched = False
+
+                    for title_id, link_count, total_link_count, doc_count in self.data_trie[prefix]:
+                        mention = Mention(self.title_trie.restore_key(title_id), prefix, start, end, link_count,
+                                          total_link_count, doc_count)
+                        ret.append(mention)
+                        matched = True
+
+                    if matched:
+                        cur = end
+                        break
+        return ret
+
+    @staticmethod
+    def build(dump_db, tokenizer, out_file, min_link_prob, min_prior_prob, min_link_count, max_mention_length,
+              pool_size, chunk_size):
+        name_dict = defaultdict(Counter)
+
+        logger.info('Iteration 1/2: Extracting all entity names...')
+        with tqdm(total=dump_db.page_size(), mininterval=0.5) as pbar:
+            initargs = (dump_db, tokenizer, max_mention_length)
+            with closing(Pool(pool_size, initializer=EntityLinker._initialize_worker, initargs=initargs)) as pool:
+                for ret in pool.imap_unordered(EntityLinker._extract_name_entity_pairs, dump_db.titles(),
+                                               chunksize=chunk_size):
+                    for text, title in ret:
+                        name_dict[text][title] += 1
+                    pbar.update()
+
+        name_counter = Counter()
+
+        disambi_matcher = re.compile(r'\s\(.*\)$')
+        for title in dump_db.titles():
+            text = normalize_text(disambi_matcher.sub('', title))
[…]
