### 2019-09-09 011812e
update README and hyperparameters
