### 2019-09-09 a93cc9a
minor fixes  on README of text classification example
