### 2019-09-12 01a088c
update the README of text classification example
