### 2019-09-19 30c13f9
add stopword removal
