### 2019-09-20 5b6035d
fix r8 dataset reader for better reproducibility
