### 2019-09-20 603b037
fix commands.md
