+++ docs/commands.md
@@ -98,7 +98,7 @@ The *build-dictionary* command builds a dictionary of words and entities.
 The *build-link-graph* command generates a sparse matrix representing the link structure between Wikipedia entities.
 
 ```text
-% wikipedia2vec build_link_graph DUMP_DB_FILE DIC_FILE OUT_FILE
+% wikipedia2vec build-link-graph DUMP_DB_FILE DIC_FILE OUT_FILE
 ```
 
 **Arguments:**
@@ -144,8 +144,8 @@ The *train-embedding* command runs the training of the embeddings.
 
 **Arguments:**
 
-- *DUMP_DB_FILE*: The database file generated using the *build\_dump\_db* command
-- *DIC_FILE*: The dictionary file generated by the *build\_dictionary* command
+- *DUMP_DB_FILE*: The database file generated using the *build-dump-db* command
+- *DIC_FILE*: The dictionary file generated by the *build-dictionary* command
 - *OUT_FILE*: The output file
 
 **Options:**
