### 2019-09-20 36a4e4b
remove unused code
