### 2019-09-21 cb7caee
update hyper parameters for r8 dataset
