### 2020-02-01 #1558 @arfon
@ionlights @usethedata - I've been in touch with the author (@ikuyamada) and they have decided to withdraw this paper from JOSS. Thanks for your efforts on this, and I'm sorry this paper didn't make it to accept/publish.
